Ays container (#35)

+++ templates/filesystem/schema.capnp
@@ -2,8 +2,8 @@
 
 struct Schema {
     storagePool @0 :Text; # Name of the parent
-    sizeOnDisk @1 :Uint64; # Amount of MiB of storage used by the filesystem
+    sizeOnDisk @1 :UInt32; # Amount of MiB of storage used by the filesystem
     readOnly @2: Bool;
-    quota @3: Uint64; # Amount of MiB that can be written to the filesystem. 0 means no quota is set.
+    quota @3: UInt32; # Amount of MiB that can be written to the filesystem. 0 means no quota is set.
     mountpoint @4:Text;
 }
